+++ internal/pkg/delivery/grpc/payment-service/transformations.go
@@ -11,12 +11,12 @@ func PaymentToProto(p *models.Payment) *proto.Payment {
 		PaymentUuid: &commonProto.UUID{Value: p.PaymentUuid.String()},
 		UserUuid:    &commonProto.UUID{Value: p.UserUuid.String()},
 		Status:      string(p.Status),
-		Price:       int64(p.Price),
+		Price:       p.Price,
 		TimeUpdated: p.TimeUpdated.Unix(),
 	}
 }
 
-func ProtoToCreatePaymentRequest(p *proto.CreatePaymentRequest) (uuid string, value int64) {
+func ProtoToCreatePaymentRequest(p *proto.CreatePaymentRequest) (uuid string, value float32) {
 	if p == nil {
 		return
 	}
